Use long literals
This patch replaces int literals that were cast to longs (e.g.,
"(long) 1)" with long literals (e.g., "1L"), making the code cleaner
and easier to maintain.

+++ src/test/java/org/apache/commons/lang3/ArrayUtilsRemoveTest.java
@@ -447,18 +447,18 @@ public class ArrayUtilsRemoveTest {
     @SuppressWarnings("cast")
     public void testRemoveElementLongArray() {
         long[] array;
-        array = ArrayUtils.removeElement((long[]) null, (long) 1);
+        array = ArrayUtils.removeElement((long[]) null, 1L);
         assertNull(array);
-        array = ArrayUtils.removeElement(ArrayUtils.EMPTY_LONG_ARRAY, (long) 1);
+        array = ArrayUtils.removeElement(ArrayUtils.EMPTY_LONG_ARRAY, 1L);
         assertTrue(Arrays.equals(ArrayUtils.EMPTY_LONG_ARRAY, array));
         assertEquals(Long.TYPE, array.getClass().getComponentType());
-        array = ArrayUtils.removeElement(new long[] {1}, (long) 1);
+        array = ArrayUtils.removeElement(new long[] {1}, 1L);
         assertTrue(Arrays.equals(ArrayUtils.EMPTY_LONG_ARRAY, array));
         assertEquals(Long.TYPE, array.getClass().getComponentType());
-        array = ArrayUtils.removeElement(new long[] {1, 2}, (long) 1);
+        array = ArrayUtils.removeElement(new long[] {1, 2}, 1L);
         assertTrue(Arrays.equals(new long[] {2}, array));
         assertEquals(Long.TYPE, array.getClass().getComponentType());
-        array = ArrayUtils.removeElement(new long[] {1, 2, 1}, (long) 1);
+        array = ArrayUtils.removeElement(new long[] {1, 2, 1}, 1L);
         assertTrue(Arrays.equals(new long[] {2, 1}, array));
         assertEquals(Long.TYPE, array.getClass().getComponentType());
     }
