+++ src/test/java/org/apache/commons/lang3/RangeTest.java
@@ -54,7 +54,7 @@ public class RangeTest {
         byteRange3 = Range.between((byte) 0, (byte) 10);
 
         intRange = Range.between(10, 20);
-        longRange = Range.between((long) 10, (long) 20);
+        longRange = Range.between(10L, 20L);
         floatRange = Range.between((float) 10, (float) 20);
         doubleRange = Range.between((double) 10, (double) 20);
     }
